% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooglevideo_auth.R
\name{service_auth}
\alias{service_auth}
\title{Service Account Auth For Google Video Intelligence API}
\usage{
service_auth(client_id, client_secret, service_json)
}
\arguments{
\item{client_id}{A string containing the client_id given from Google.}

\item{client_secret}{A string containin the client secret given from Google.}

\item{service_json}{A string path to your service account .json file.}
}
\value{
A service token for use within the session.
}
\description{
Ideal for use in automated scripts or Shiny apps. No
browser will pop up during this method. Instead, you will
need to receive a service account `.json` file from
Google. This file will contain your private key, which
should be kept secret. Additionally, you will need
OAuth creds to obtain the client id and secret.
}
