% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_b64}
\alias{to_b64}
\title{Convert To Base64 Encoded String}
\usage{
to_b64(asset_path)
}
\arguments{
\item{asset_path}{A string consisting of a file path
to the asset to convert (.mp4, .gif, etc..).}
}
\value{
A base64 encoded string.
}
\description{
Convert To Base64 Encoded String
}
