% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooglevideo_auth.R
\name{web_auth}
\alias{web_auth}
\title{Web Browser Auth For Google Video Intelligence API}
\usage{
web_auth(client_id, client_secret)
}
\arguments{
\item{client_id}{A string containing the client_id given from Google.}

\item{client_secret}{A string containing the client_secret give from Google.}
}
\value{
An access token for use within the session.
}
\description{
This method is best for interactive sessions. You
will need to register a project with Google and
obtain OAuth credentials.
}
